Add basic Play Integrity to TrivialKart-unity
* Updated README for public release

* Bump bundle version to 1.1 for upcoming public release.

* Implements an integrity check the first time the store is
opened during a play session. A summary string of the verdict
results is displayed on the store page.

Server code is intentionally not included. The README has been updated
to define the HTTP endpoints that TrivialKart expects, with
links to the relevant DAC guides. The developer is expected to
implement using their preferred platform.

Once the Play Integrity codelab is published, the readme will be
updated to point to it as an example (compatible) server implementation.

Change-Id: Ia5c805d8d892854e9f495d26e3f80563986d7a05