AGDKTunnel Vulkan support CL 1 of 3
Support CL for AGDKTunnel Vulkan support. This CL includes
a minimal integration of BaseGameFramework to add the
DisplayManager API, which is used to abstract graphics API
initialization and swapchain init/present. Instead of calling
EGL/GL directly, AGDKTunnel now uses DisplayManager, so it will
be able to select between GLES or Vulkan.

Only includes the GLES backend for DisplayManager, Vulkan backend
will be added in CL 2 of 3.

Includes a few minor modifications to SimpleRenderer to use
Display Manager, and the upcoming Vulkan implementation CL.

Test: Build and run sample, game should function the same.
Change-Id: Ia334ef686b69f9de1ccb3d0683cc16782d312446