Final AGDKTunnel BGF migration (2/3)
This CL series adds the rest of the BaseGameFramework and refactors
AGDKTunnel to use it instead of NDK/ADGK directly.

This CL adds the source files for the Androd implementation of the
framework init, filesystem manager, system event manager,
user input manager and Android utility java class.

Test: Build and run sample. This is just a implementation review CL,
integration code and adding the rest of BGF to the bulid is in the
next CL

Change-Id: I95a6c3771235593d615d4a6a7f81221de420b1b3