Final AGDKTunnel BGF migration (3/3)
This CL series adds the rest of the BaseGameFramework and refactors
AGDKTunnel to use it instead of NDK/ADGK directly.

This CL refactors the native_engine game file to use the BGF managers,
and splits it into native_engine/tunnel_engine. The native_engine is
intended to be a base class common between samples and will be
moved into common/src when updating the other samples.

Game code updated to call TunnelEngine::GetInstance instead
of NativeEngine.

Also updates build script versions for NDK, AGP and AGDK libraries.

Test: Build and run in Android Studio
Change-Id: I6d2b7d00028e4e6fad5f91279f79f924485485b6